FIX: NPE during test case

+++ src/main/java/io/ebeaninternal/dbmigration/ddlgeneration/platform/BaseTableDdl.java
@@ -782,7 +782,7 @@ public class BaseTableDdl implements TableDdl {
 
     alterTableDropColumn(writer.apply(), tableName, dropColumn.getColumnName());
     MTable table = writer.getTable(tableName);
-    if (table.isWithHistory() && historySupport == HistorySupport.TRIGGER_BASED) {
+    if (table != null && table.isWithHistory() && historySupport == HistorySupport.TRIGGER_BASED) {
       // also drop from the history table
       regenerateHistoryTriggers(tableName, HistoryTableUpdate.Change.DROP, dropColumn.getColumnName());
       alterTableDropColumn(writer.apply(), historyTable(tableName), dropColumn.getColumnName());
